kernel: move rdtsc patches relative into its own source file.
