+++ CMakeLists.txt
@@ -184,6 +184,7 @@ file (GLOB CXBXR_HEADER_EMU
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuFile.h"
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuFS.h"
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuNtDll.h"
+ "${CXBXR_ROOT_DIR}/src/core/kernel/support/PatchRdtsc.hpp"
  "${CXBXR_ROOT_DIR}/src/devices/ADM1032Device.h"
  "${CXBXR_ROOT_DIR}/src/devices/EEPROMDevice.h"
  "${CXBXR_ROOT_DIR}/src/devices/network/NVNetDevice.h"
@@ -361,6 +362,7 @@ file (GLOB CXBXR_SOURCE_EMU
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuFile.cpp"
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuFS.cpp"
  "${CXBXR_ROOT_DIR}/src/core/kernel/support/EmuNtDll.cpp"
+ "${CXBXR_ROOT_DIR}/src/core/kernel/support/PatchRdtsc.cpp"
  "${CXBXR_ROOT_DIR}/src/devices/ADM1032Device.cpp"
  "${CXBXR_ROOT_DIR}/src/devices/EEPROMDevice.cpp"
  "${CXBXR_ROOT_DIR}/src/devices/network/NVNetDevice.cpp"
